requirements.txt
instabot==0.117.0
Pillow==7.2.0
python-dotenv==0.10.3
requests==2.24.0
